pyproject.toml
[tool.poetry]
name = "progaming-test-task"
version = "0.1.0"
description = ""
authors = ["Andrii Hermak <hermak20102004@gmail.com>"]
readme = "README.md"
packages = [{include = "progaming_test_task"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
psycopg = "^3.2.2"
psycopg-binary = "^3.2.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
django-redis = "^5.4.0"
redis = "^5.0.8"
django-debug-toolbar = "^4.4.6"
channels = {extras = ["daphne"], version = "^4.1.0"}
channels-redis = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
